+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.10.1] - 2026-03-013
+
+### Enhancement
+- 🔧 优化：优化算法，提升GanttLinks以及GanttConflicts的绘制性能
+- 🔧 Optimized: Algorithm optimizations for improved GanttLinks and GanttConflicts rendering performance
+
 ## [1.10.0] - 2026-03-07
 
 ### Fixed